Resolves requested changes
