+++ basyx.conceptdescriptionrepository/basyx.conceptdescriptionrepository-http/src/main/java/org/eclipse/digitaltwin/basyx/conceptdescriptionrepository/http/factory/FilterConceptDescriptionFactory.java
@@ -1,0 +1,75 @@
+/*******************************************************************************
+ * Copyright (C) 2023 the Eclipse BaSyx Authors
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * 
+ * SPDX-License-Identifier: MIT
+ ******************************************************************************/
+
+package org.eclipse.digitaltwin.basyx.conceptdescriptionrepository.http.factory;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.eclipse.digitaltwin.aas4j.v3.model.ConceptDescription;
+import org.eclipse.digitaltwin.aas4j.v3.model.Reference;
+import org.eclipse.digitaltwin.basyx.conceptdescriptionrepository.ConceptDescriptionRepository;
+
+/**
+ * Factory class to create filtered ConceptDescriptions
+ * 
+ * @author danish
+ *
+ */
+public class FilterConceptDescriptionFactory {
+
+	private ConceptDescriptionRepository repository;
+	private String idShort;
+	private Reference isCaseOf;
+	private Reference dataSpecificationRef;
+
+	public FilterConceptDescriptionFactory(ConceptDescriptionRepository conceptDescriptionRepository,
+			String idShort, Reference isCaseOf,
+			Reference dataSpecificationRef) {
+		super();
+		this.repository = conceptDescriptionRepository;
+		this.idShort = idShort;
+		this.isCaseOf = isCaseOf;
+		this.dataSpecificationRef = dataSpecificationRef;
+	}
+	
+	/**
+	 * Filters ConceptDescriptions from repository
+	 * 
+	 * @apiNote Atleast one attribute should be set
+	 * 
+	 * @return a filtered collection of ConceptDescriptions
+	 */
+	public List<ConceptDescription> create() {
+		if (idShort != null)
+			return repository.getAllConceptDescriptionsByIdShort(idShort).stream().collect(Collectors.toList());
+			
+		if (isCaseOf != null)
+			return repository.getAllConceptDescriptionsByIsCaseOf(isCaseOf).stream().collect(Collectors.toList());
+		
+		return repository.getAllConceptDescriptionsByDataSpecificationReference(dataSpecificationRef).stream().collect(Collectors.toList());
+	}
+
+}
